shop/requirements.txt
Django==3.1.7
django-mptt==0.12.0
psycopg2==2.8.6
pillow==8.2.0
python-dotenv==0.17.1
